Working on message fixes / chatting. Not ready for merge

+++ scaffold/src/Classes/CityVill.h
@@ -9,7 +9,7 @@
 #import "SpecialIncomeCounters.h"
 
 @interface CityVill : SpecialIncomeCounters{
-    int *_combatValue;
+    NSInteger *_combatValue;
     //NSString *_fileName;
 }
